+++ src/main/java/baseball/Application.java
@@ -2,9 +2,9 @@ package baseball;
 
 import baseball.app.ball.generator.AutoBallsGenerator;
 import baseball.app.ball.generator.ManualBallsGenerator;
-import baseball.app.game.Computer;
+import baseball.app.participant.Computer;
 import baseball.app.game.Game;
-import baseball.app.game.Player;
+import baseball.app.participant.Player;
 
 public class Application {
     public static void main(String[] args) {
