+++ src/main/java/baseball/app/ball/generator/ManualBallsGenerator.java
@@ -6,7 +6,7 @@ import camp.nextstep.edu.missionutils.Console;
 import java.util.ArrayList;
 import java.util.List;
 
-import static baseball.app.constants.GameConstants.MAX_BALL_COUNT;
+import static baseball.app.ball.constants.GameConstants.MAX_BALL_COUNT;
 
 public class ManualBallsGenerator extends AbstractBallsGenerator {
