+++ src/main/java/baseball/app/participant/Player.java
@@ -1,7 +1,8 @@
-package baseball.app.game;
+package baseball.app.participant;
 
 import baseball.app.ball.Balls;
 import baseball.app.ball.generator.BallsGenerator;
+import baseball.app.game.GameMenu;
 import baseball.app.view.GameView;
 import camp.nextstep.edu.missionutils.Console;
